refactor: moves "utilities/Repository.ts" to "utilities/Repository/definition.ts"
